requirements.txt
openai
python-dotenv
langchain
streamlit
crewai
crewai-tools